Move graph type
