+++ src/interactive.rs
@@ -1,6 +1,7 @@
 use super::config::{get_config, Config};
+use super::graph::GraphType;
 use super::interface::{DaemonControl, DaemonController, Get, Getter, Interface, Set, Setter};
-use super::settings::{GraphType, Settings};
+use super::settings::Settings;
 use colored::Colorize;
 use std::io::{stdin, stdout, Write};
